DB Remote Crash bug corrected? - (NSInteger)read:(uint8_t *)buffer maxLength:(NSUInteger)len can return -1 : you cannot store result in unsigned int...
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@8916 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ nitrogen/Sources/N2Connection.mm
@@ -147,7 +147,7 @@ NSString* N2ConnectionStatusDidChangeNotification = @"N2ConnectionStatusDidChang
 		_hasBytesAvailable = NO;
 		
 		NSUInteger maxLength = 1024; uint8_t buffer[maxLength];
-		unsigned int length = [_inputStream read:buffer maxLength:maxLength];
+		NSInteger length = [_inputStream read:buffer maxLength:maxLength];
 		if (length > 0) {
 			DLog(@"%@ Read %d Bytes", self, length);
 			//			std::cerr << [[NSString stringWithFormat:@"%@ Read %d Bytes", self, length] UTF8String] << ": ";
